requirements.txt
PyYAML
dateparser
tqdm
beautifulsoup4
Pyzotero
thefuzz